backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
alembic
PyJWT
passlib[argon2]
